refactor: 7711 - remove the SNS package
