+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/Configuration/AwsSnsConfiguration.cs
@@ -1,9 +1,8 @@
 using xxENSONOxx.xxSTACKSxx.Shared.Configuration;
 
-namespace xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS;
+namespace xxENSONOxx.xxSTACKSxx.Infrastructure.Configuration;
 
 public class AwsSnsConfiguration
 {
     public Secret TopicArn { get; init; } = null!;
 }
-
