--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS/Events/Events.cs
@@ -1,10 +1,0 @@
-namespace xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS.Events;
-
-public enum EventCode
-{
-    GeneralException = 123456899,
-
-    PublishEventRequested = 123456801,
-    PublishEventCompleted = 123456802,
-    PublishEventFailed = 123456803,
-}
