--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS/xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <Description>AWS SNS package for common features used throughout the projects.</Description>
-        <Company>Ensono</Company>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageProjectUrl>https://github.com/ensono/stacks-dotnet-packages-sns</PackageProjectUrl>
-        <RepositoryUrl>https://github.com/ensono/stacks-dotnet-packages-sns</RepositoryUrl>
-        <RepositoryType>GIT</RepositoryType>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.301" />
-      <PackageReference Include="AWSSDK.SimpleNotificationService" Version="3.7.400.5" />
-      <PackageReference Include="Microsoft.Extensions.Options" Version="8.0.2" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions.csproj" />
-      <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Shared.Configuration\xxENSONOxx.xxSTACKSxx.Shared.Configuration.csproj" />
-      <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Shared.DependencyInjection\xxENSONOxx.xxSTACKSxx.Shared.DependencyInjection.csproj" />
-    </ItemGroup>
-
-</Project>
