+++ xxENSONOxx.xxSTACKSxx.sln
@@ -75,10 +75,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests.csproj", "{688708D6-CF24-452C-BBFD-59161F982AA0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS\xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS.csproj", "{7BEFAEE8-4F24-434B-ABEA-4E904804266C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS.Tests\xxENSONOxx.xxSTACKSxx.Shared.Messaging.AWS.SNS.Tests.csproj", "{C63011EA-2C53-495B-96F8-71124DB6236C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub.csproj", "{52FCB145-BC6E-4B50-AE2A-C51C3466EB92}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub.Tests\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub.Tests.csproj", "{70705814-2618-48A3-AD93-EF77089BE7B3}"
@@ -565,30 +561,6 @@ Global
 		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x64.Build.0 = Release|Any CPU
 		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x86.ActiveCfg = Release|Any CPU
 		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x86.Build.0 = Release|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|x64.Build.0 = Debug|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|x86.Build.0 = Debug|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Release|x64.ActiveCfg = Release|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Release|x64.Build.0 = Release|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Release|x86.ActiveCfg = Release|Any CPU
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Release|x86.Build.0 = Release|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Debug|x64.Build.0 = Debug|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Debug|x86.Build.0 = Debug|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Release|x64.ActiveCfg = Release|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Release|x64.Build.0 = Release|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Release|x86.ActiveCfg = Release|Any CPU
-		{C63011EA-2C53-495B-96F8-71124DB6236C}.Release|x86.Build.0 = Release|Any CPU
 		{52FCB145-BC6E-4B50-AE2A-C51C3466EB92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{52FCB145-BC6E-4B50-AE2A-C51C3466EB92}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{52FCB145-BC6E-4B50-AE2A-C51C3466EB92}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -881,8 +853,6 @@ Global
 		{7DD7AD0F-07EE-438F-8032-B0BB40F26E85} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{688708D6-CF24-452C-BBFD-59161F982AA0} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{7BEFAEE8-4F24-434B-ABEA-4E904804266C} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{C63011EA-2C53-495B-96F8-71124DB6236C} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{52FCB145-BC6E-4B50-AE2A-C51C3466EB92} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{70705814-2618-48A3-AD93-EF77089BE7B3} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
